+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added license to test files.
 - Fixed unterminated comment.
 - Fixed BLE and WiFI updates.
+- Improving use of pTab4Pwr.
 
 ### Hardware